fix: address PR feedback and apply correct namespace
- Change module namespaces and package imports to `com.ram.pocketserver`.
- Rename app in resources to `pocket server`.
- Set HardwareDetector fallback to `false` for `hasGpuDelegate`.
- Configure matching `compileSdk` and `targetSdk` globally via versions catalog.
- Add stub `proguard-rules.pro` for release builds.
- Refactor `seed` param in `InferenceEngine` from Long to Int.
- Refine Hilt `@Singleton` scoping for `AppModule`.
- Clean up unused Kotlin coroutines imports in `LiteRTEngine`.
- Add explicit Flow `collect` import in `MainActivity`.

+++ inference/inference-api/src/main/java/com/ram/pocketserver/inference/InferenceEngine.kt
@@ -10,7 +10,7 @@ data class ModelLoadConfig(
     val contextLength: Int = 4096,
     val topK: Int = 40,
     val temperature: Float = 0.7f,
-    val seed: Long = -1L,
+    val seed: Int = -1,
 )
 
 data class ChatMessage(val role: String, val content: String)
@@ -21,7 +21,7 @@ data class GenerationParams(
     val topP: Float = 0.9f,
     val topK: Int = 40,
     val repeatPenalty: Float = 1.1f,
-    val seed: Long = -1L,
+    val seed: Int = -1,
     val stop: List<String> = emptyList(),
     val stream: Boolean = true,
     val minP: Float = 0.05f,
